Accept a `--port` or `-p` argument for `testdrb`
Since `spork` is able to accept a port argument, testdrb should be able to as
well.

This became important today because I am running spork on the same server as a
coworker and we needed to be able to run on different ports. `testdrb` could not
listen on any port other than the default, so I added functionality to deal with
that.